pyproject.toml
[project]
name = "overlay"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "json5>=0.12.0",
    "psutil>=7.0.0",
    "pyinstaller>=6.13.0",
    "pywin32>=310; sys_platform == 'win32'",
    "toml>=0.10.2",
    "evdev; sys_platform == 'linux'",
    "keyboard>=0.13.5",
    "requests>=2.32.3",
    "pyyaml>=6.0.2",
    "attrs>=25.3.0",
    "websockets>=15.0.1",
    "pycaw>=20240210; sys_platform == 'win32'",
    "markupsafe>=3.0.2",
    "typer>=0.16.0",
    "numba>=0.61.2",
    "textual>=3.5.0",
    "pulsectl>=24.12.0; sys_platform == 'linux'",
    "pillow>=11.3.0",
    "opencv-python>=4.12.0.88",
    "pycv3>=1.2.0",
    "fs>=2.4.16",
    "jinja2>=3.1.6",
    "pyside6==6.10.1",
    "setuptools<81",
    "pulsectl>=24.12.0; sys_platform == 'linux'",
    "pydantic>=2.11.7",
    "pyi18n-v2>=1.2.2",
    "methodtools>=0.4.7",
]
